requirements.txt
PyGetWindow~=0.0.9
numpy~=1.20.2
opencv-python~=4.5.1.48
waiting~=1.4.1
Pillow~=8.2.0
pywin32~=300